+++ packages/modern-browsers/package.js
@@ -1,6 +1,6 @@
 Package.describe({
   name: 'modern-browsers',
-  version: '0.1.8-beta272.0',
+  version: '0.1.8-rc272.0',
   summary:
     'API for defining the boundary between modern and legacy ' +
     'JavaScript clients',