+++ tools/cli/commands.js
@@ -514,7 +514,8 @@ export const AVAILABLE_SKELETONS = [
   DEFAULT_SKELETON,
   "typescript",
   "vue",
-  "svelte"
+  "svelte",
+  "tailwind",
 ];
 
 main.registerCommand({
@@ -533,6 +534,7 @@ main.registerCommand({
     typescript: { type: Boolean },
     apollo: { type: Boolean },
     svelte: { type: Boolean },
+    tailwind: { type: Boolean },
   },
   catalogRefresh: new catalog.Refresh.Never()
 }, function (options) {
@@ -904,6 +906,7 @@ main.registerCommand({
     cmd("meteor create --svelte     # to create a basic Svelte app");
     cmd("meteor create --typescript # to create an app using TypeScript and React");
     cmd("meteor create --blaze      # to create an app using Blaze");
+    cmd("meteor create --tailwind   # to create an app using React and Tailwind");
   }
 
   Console.info("");